requirements.txt
# Core Machine Learning
torch>=2.4.0
transformers>=4.45.0
peft>=0.12.0
datasets>=3.0.0
sentence-transformers>=3.1.0

# Data Processing & Math
pandas>=2.2.0
scikit-learn>=1.5.0
numpy>=1.26.0

# API & Utilities
openai>=1.40.0
python-dotenv>=1.0.1
tqdm>=4.66.0